+++ legends/ui/session.py
@@ -436,11 +436,12 @@ class SurvivalEffects(ModalMessage):
         """
         saveslot = self.root.session.saveslot
         count = 0
-        for effect, duration in saveslot.survivalEffects.items():
-            tk.Label(
-                master, text=effect
-            ).grid(row=count, column=0, sticky=tk.W, padx=(20,0))
-            tk.Label(
-                master, text='({} battles remaining)'.format(duration)
-            ).grid(row=count, column=1, sticky=tk.E, padx=(0,20))
-            count += 1
+        for effect, durations in saveslot.survivalEffects.items():
+            for duration in durations:
+                tk.Label(
+                    master, text=effect
+                ).grid(row=count, column=0, sticky=tk.W, padx=(20,0))
+                tk.Label(
+                    master, text='({} battles remaining)'.format(duration)
+                ).grid(row=count, column=1, sticky=tk.E, padx=(0,20))
+                count += 1